I tried to reduce the endless number of compiler-warnings: Boost switched from int to size_t for specifying the vector/matrix dimensions. Unfortunately Lapack(Fortran) still expects ints for all the values, so I added some casts to the bindings to stop this mess. I hope, that should do it.
You can see clearly now the warnings coming from our own code again :)

+++ 3rd/boost-bindings/boost/numeric/bindings/traits/detail/ublas_ordering.hpp
@@ -33,12 +33,12 @@ namespace boost { namespace numeric { namespace bindings { namespace traits {
 
       template <typename M>
       static int leading_dimension( M const& m ) {
-        return m.size2() ;
+        return static_cast < int > ( m.size2() );
       }
 
       template <typename M>
       static int stride1( M const& m ) {
-        return m.size2() ;
+        return static_cast < int > ( m.size2() );
       }
 
       template <typename M>
@@ -54,7 +54,7 @@ namespace boost { namespace numeric { namespace bindings { namespace traits {
 
       template <typename M>
       static int leading_dimension( M const& m ) {
-        return m.size1() ;
+        return static_cast < int > ( m.size1() );
       }
 
       template <typename M>
@@ -64,7 +64,7 @@ namespace boost { namespace numeric { namespace bindings { namespace traits {
 
       template <typename M>
       static int stride2( M const& m ) {
-        return m.size1() ;
+        return static_cast < int > ( m.size1() );
       }
     };
